Write size of message... twice!
The reason is: ProtoMessage.writeTo reads all contents of input stream, so we have to use ProtoMessage.writeDelimitedTo, which in case writes size of message. Also we have to write message size first for non-blocking architecture to work. Otherwise we can't read size of message and then .parseDelimitedFrom. Anyway, latter is a blocking method